+++ src/components/layout/StatusBar.tsx
@@ -18,7 +18,7 @@ import {
   isIntuneWorkspace,
   useUiStore,
 } from "../../stores/ui-store";
-import { useIntuneStore } from "../../stores/intune-store";
+import { useIntuneStore } from "../../workspaces/intune/intune-store";
 import { useDsregcmdStore } from "../../workspaces/dsregcmd/dsregcmd-store";
 import { useDeploymentStore } from "../../workspaces/deployment/deployment-store";
 import { useSysmonStore } from "../../workspaces/sysmon/sysmon-store";
